feat(server): move dto to module directory
