+++ server/src/gates/dtos/requests/activate-device-request.dto.ts
