--- server/src/mqtt/dtos/requests/activate-mqtt-device-request.dto.ts
@@ -1,7 +1,0 @@
-import { IsString } from 'class-validator';
-import { ActivateDeviceRequestDto } from 'src/common/dtos/requests/activate-device-request.dto';
-
-export class ActivateMqttDeviceRequestDto extends ActivateDeviceRequestDto {
-  @IsString()
-  topic: string;
-}
